+++ packages/breadcrumb/src/style.tsx
@@ -3,19 +3,19 @@ import { globalColor, illaPrefix } from "@illa-design/theme"
 
 export const itemCss = css`
   display: inline-block;
-  color: ${globalColor(`--${illaPrefix}-gray-04`)};
+  color: ${globalColor(`--${illaPrefix}-grayBlue-04`)};
   padding: 4px;
   border-radius: 4px;
   transition: all 0.2s;
 `
 
 export const itemCurrentCss = css`
-  color: ${globalColor(`--${illaPrefix}-gray-02`)};
+  color: ${globalColor(`--${illaPrefix}-grayBlue-02`)};
 `
 
 export const itemHoverCss = css`
   &:hover {
-    background-color: ${globalColor(`--${illaPrefix}-gray-09`)};
+    background-color: ${globalColor(`--${illaPrefix}-grayBlue-09`)};
     color: ${globalColor(`--${illaPrefix}-blue-03`)};
     cursor: pointer;
   }
@@ -23,7 +23,7 @@ export const itemHoverCss = css`
 
 export const separatorCss = css`
   margin: 0 12px;
-  color: ${globalColor(`--${illaPrefix}-gray-06`)};
+  color: ${globalColor(`--${illaPrefix}-grayBlue-06`)};
   position: relative;
   top: 1.5px;
   transform: scale(0.7);