+++ packages/menu/src/styles/item-group.tsx
@@ -7,8 +7,8 @@ import { applyPopButtonCss } from "../style"
 export function applyItemGroupCss(theme: Theme = "light"): SerializedStyles {
   const color =
     theme === "dark"
-      ? globalColor(`--${illaPrefix}-gray-04`)
-      : globalColor(`--${illaPrefix}-gray-04`)
+      ? globalColor(`--${illaPrefix}-grayBlue-04`)
+      : globalColor(`--${illaPrefix}-grayBlue-04`)
   return css`
     color: ${color};
     font-size: 14px;