+++ packages/select/src/select-view.tsx
@@ -104,11 +104,6 @@ export const SelectView = forwardRef<HTMLDivElement, SelectViewProps>(
       }
     }, [popupVisible])
 
-    function forceUpdate() {
-      const [, dispatch] = useReducer((v) => v + 1, 0)
-      return dispatch
-    }
-
     const tryTriggerFocusChange = (
       action: "focus" | "blur",
       event: SyntheticEvent,
@@ -146,7 +141,7 @@ export const SelectView = forwardRef<HTMLDivElement, SelectViewProps>(
       <LoadingIcon
         spin
         css={css`
-          color: ${globalColor(`--${illaPrefix}-gray-07`)};
+          color: ${globalColor(`--${illaPrefix}-grayBlue-07`)};
         `}
       />
     ) : showSearch && popupVisible ? (